requirements.txt
Django==5.0.1
python-decouple==3.8
whitenoise==6.6.0
